added some classes for the die rolls
I added the die parent class and child class for potion, weapon, and armor.
Right now die only holds the number of sides, but I might standardize the rolling into the die class.
As is, armor works by taking 1 evasion roll and returning -10 for that one, for the rest it returns the amount that the damage should be reduced by. Potion and weapon work by calling the heal and damage functions.